crud kuota sekolah

+++ backend/models/mhs.js
@@ -3,11 +3,11 @@ module.exports = (sequelize, Sequelize) => {
     nama_mhs: {
       type: Sequelize.STRING
     }, nim_mhs: {
-      type: Sequelize.INTEGER
+      type: Sequelize.BIGINT
     }, ipk_mhs: {
       type: Sequelize.FLOAT
     }, nokwitansi_mhs : {
-      type: Sequelize.INTEGER
+      type: Sequelize.BIGINT
     }
   });
   return mhs;
